requirements.txt
python-telegram-bot==13.9
emoji==1.2.0
Pillow==8.2.0
PyQRCode==1.2.1
pypng==0.0.20
firebase-admin==5.2.0
sib-api-v3-sdk==7.3.0
flask==2.0.2